feedback: expose unwrap on client instance
