+++ src/endpoints/CollectionLayers.js
@@ -8,7 +8,7 @@ import type {
   UpdatedCollectionLayer
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class CollectionLayers extends Endpoint {
   add(
