+++ src/endpoints/Comments.js
@@ -12,7 +12,7 @@ import type {
   RequestOptions
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Comments extends Endpoint {
   async create(
