+++ src/endpoints/Data.js
@@ -5,7 +5,7 @@ import type {
   RequestOptions
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Data extends Endpoint {
   async info(
