+++ src/endpoints/Layers.js
@@ -9,7 +9,7 @@ import type {
   RequestOptions
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Layers extends Endpoint {
   async info(
