+++ src/endpoints/Memberships.js
@@ -8,7 +8,7 @@ import type {
   RequestOptions
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Memberships extends Endpoint {
   info(
