+++ src/endpoints/Organizations.js
@@ -5,7 +5,7 @@ import type {
   RequestOptions
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Organizations extends Endpoint {
   info(
