+++ src/endpoints/Sections.js
@@ -1,7 +1,7 @@
 // @flow
 import type { OrganizationDescriptor, RequestOptions, Section } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Sections extends Endpoint {
   list(
