+++ src/endpoints/Users.js
@@ -7,7 +7,7 @@ import type {
   UserDescriptor
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Users extends Endpoint {
   info(descriptor: UserDescriptor, requestOptions: RequestOptions = {}) {
