+++ src/endpoints/Webhooks.js
@@ -11,7 +11,7 @@ import type {
   WebhookEvent
 } from "../types";
 import Endpoint from "../endpoints/Endpoint";
-import { wrap } from "../response";
+import { wrap } from "../util/helpers";
 
 export default class Users extends Endpoint {
   list(
